Move job dq back to job service.
